+++ docker/lib/puppet/parser/functions/docker_service_flags.rb
@@ -1,10 +1,12 @@
+# frozen_string_literal: true
+
 require 'shellwords'
 #
 # docker_service_flags.rb
 #
 module Puppet::Parser::Functions
   # Transforms a hash into a string of docker swarm init flags
-  newfunction(:docker_service_flags, :type => :rvalue) do |args|
+  newfunction(:docker_service_flags, type: :rvalue) do |args|
     opts = args[0] || {}
     flags = []
 
@@ -12,7 +14,7 @@ module Puppet::Parser::Functions
       flags << '--detach'
     end
 
-    if opts['service_name'].to_s != 'undef'
+    if opts['service_name'] && opts['service_name'].to_s != 'undef'
       flags << "'#{opts['service_name']}'"
     end
 
@@ -28,11 +30,27 @@ module Puppet::Parser::Functions
       end
     end
 
-    if opts['publish'].to_s != 'undef'
+    if opts['mounts'].is_a? Array
+      opts['mounts'].each do |mount|
+        flags << "--mount #{mount}"
+      end
+    end
+
+    if opts['networks'].is_a? Array
+      opts['networks'].each do |network|
+        flags << "--network #{network}"
+      end
+    end
+
+    if opts['publish'].is_a? Array
+      opts['publish'].each do |port|
+        flags << "--publish #{port}"
+      end
+    elsif opts['publish'].to_s != 'undef'
       flags << "--publish '#{opts['publish']}'"
     end
 
-    if opts['replicas'].to_s != 'undef'
+    if opts['replicas'] && opts['replicas'].to_s != 'undef'
       flags << "--replicas '#{opts['replicas']}'"
     end
 
@@ -40,11 +58,11 @@ module Puppet::Parser::Functions
       flags << '--tty'
     end
 
-    if opts['user'].to_s != 'undef'
+    if opts['user'] && opts['user'].to_s != 'undef'
       flags << "--user '#{opts['publish']}'"
     end
 
-    if opts['workdir'].to_s != 'undef'
+    if opts['workdir'] && opts['workdir'].to_s != 'undef'
       flags << "--workdir '#{opts['workdir']}'"
     end
 
@@ -54,19 +72,21 @@ module Puppet::Parser::Functions
       end
     end
 
-    if opts['host_socket'].to_s != 'undef'
+    if opts['host_socket'] && opts['host_socket'].to_s != 'undef'
       flags << "-H '#{opts['host_socket']}'"
     end
 
-    if opts['registry_mirror'].to_s != 'undef'
+    if opts['registry_mirror'] && opts['registry_mirror'].to_s != 'undef'
       flags << "--registry-mirror='#{opts['registry_mirror']}'"
     end
 
-    if opts['image'].to_s != 'undef'
+    if opts['image'] && opts['image'].to_s != 'undef'
       flags << "'#{opts['image']}'"
     end
 
-    if opts['command'].to_s != 'undef'
+    if opts['command'].is_a? Array
+      flags << opts['command'].join(' ')
+    elsif opts['command'] && opts['command'].to_s != 'undef'
       flags << opts['command'].to_s
     end