+++ docker/lib/puppet/parser/functions/docker_swarm_join_flags.rb
@@ -1,10 +1,12 @@
+# frozen_string_literal: true
+
 require 'shellwords'
 #
 # docker_swarm_join_flags.rb
 #
 module Puppet::Parser::Functions
   # Transforms a hash into a string of docker swarm init flags
-  newfunction(:docker_swarm_join_flags, :type => :rvalue) do |args|
+  newfunction(:docker_swarm_join_flags, type: :rvalue) do |args|
     opts = args[0] || {}
     flags = []
 
@@ -12,15 +14,15 @@ module Puppet::Parser::Functions
       flags << 'join'
     end
 
-    if opts['advertise_addr'].to_s != 'undef'
+    if opts['advertise_addr'] && opts['advertise_addr'].to_s != 'undef'
       flags << "--advertise-addr '#{opts['advertise_addr']}'"
     end
 
-    if opts['listen_addr'].to_s != 'undef'
+    if opts['listen_addr'] && opts['listen_addr'].to_s != 'undef'
       flags << "--listen-addr \"#{opts['listen_addr']}\""
     end
 
-    if opts['token'].to_s != 'undef'
+    if opts['token'] && opts['token'].to_s != 'undef'
       flags << "--token '#{opts['token']}'"
     end